requirements.txt
acres==0.2.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
bcrypt==3.1.7
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
Django==5.1.5
django-cors-headers==4.7.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
djangorestframework==3.15.2
etelemetry==0.3.1
fastapi==0.115.6
filelock==3.16.1
fitz==0.0.1.dev2
frontend==0.0.3
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
looseversion==1.3.0
lxml==5.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pillow==11.1.0
prov==2.0.1
psycopg2==2.9.10
puremagic==1.28
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydot==3.0.4
PyMuPDF==1.25.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.15.1
selenium==4.27.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.3
traits==6.4.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
wsproto==1.2.0
